Collected keys are now tracked between saving/loading.
Collected keys are represented as a single string in an xml attribute
named "keys". Each character of the string represents one key. for
example, if the player had collected all keys it may look like this:
"rbyg".

The load function iterates through each of these characters and adds
those characters as keys to the new game object.